move search results aggregations context
