+++ src/components/SearchForm/SearchForm.tsx
@@ -10,7 +10,7 @@ import { BASE_URL, PATHS } from "../../config/constants"
 import EDSLink from "../EDSLink"
 import useLoading from "../../hooks/useLoading"
 import RefineSearch from "../RefineSearch/RefineSearch"
-import { SearchResultsAggregationsContext } from "../../../pages/search/SearchResultsAggregationsContext"
+import { SearchResultsAggregationsContext } from "../../context/SearchResultsAggregationsContext"
 import type { Aggregation } from "../../types/filterTypes"
 import AppliedFilters from "../SearchFilters/AppliedFilters"
 import { collapseMultiValueQueryParams } from "../../utils/refineSearchUtils"
